Try to reduce some code duplication for sonar.
